# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing and querying time series data from IoT sensors\n\nThis API allows you to upload CSV files containing time series data,\nquery the data with time range filters, and manage datasources.\n\nSupported timestamp formats: ISO8601, RFC3339, Unix timestamps (seconds/milliseconds), Julian Day",
        "title": "IoT Data Sandbox API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/layers/{id}": {
            "get": {
                "description": "Get a specific layer by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "layers"
                ],
                "summary": "Get layer details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.LayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a layer by ID",
                "tags": [
                    "layers"
                ],
                "summary": "Delete a layer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/layers/{id}/color": {
            "put": {
                "description": "Update the display color of a layer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "layers"
                ],
                "summary": "Update layer color",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Color value (hex format)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.UpdateColorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.LayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/layers/{id}/data": {
            "get": {
                "description": "Get the time series data points for a layer (includes data from associated datasource)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "layers"
                ],
                "summary": "Get layer time series data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start time in RFC3339 format (e.g., 2024-01-01T00:00:00Z)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time in RFC3339 format (e.g., 2024-01-01T23:59:59Z)",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.DataQueryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/layers/{id}/display-window": {
            "put": {
                "description": "Update the display time window (start and end times) for a layer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "layers"
                ],
                "summary": "Update layer display time window",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Display window times",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.UpdateDisplayWindowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.LayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/layers/{id}/duplicate": {
            "post": {
                "description": "Create a copy of a layer with a new name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "layers"
                ],
                "summary": "Duplicate a layer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New layer name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.DuplicateLayerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/data.LayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/layers/{id}/load-csv": {
            "post": {
                "description": "Load data from an existing CSV file into the layer (creates a new datasource)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "layers"
                ],
                "summary": "Load CSV data into layer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "CSV filename",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.LoadCSVRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.LayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/layers/{id}/visibility": {
            "put": {
                "description": "Update the visibility state of a layer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "layers"
                ],
                "summary": "Update layer visibility",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Visibility state",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.UpdateVisibilityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.LayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.DataPoint": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "data.DataQueryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.DataPoint"
                    }
                },
                "end_time": {
                    "type": "string"
                },
                "row_count": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
        "data.DuplicateLayerRequest": {
            "type": "object",
            "properties": {
                "new_name": {
                    "type": "string"
                }
            }
        },
        "data.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "data.LayerResponse": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "data_layer_id": {
                    "type": "integer"
                },
                "data_source_id": {
                    "type": "integer"
                },
                "display_end_time": {
                    "type": "string"
                },
                "display_start_time": {
                    "type": "string"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "project_id": {
                    "type": "integer"
                },
                "z_index": {
                    "type": "integer"
                }
            }
        },
        "data.LoadCSVRequest": {
            "type": "object",
            "properties": {
                "csv_filename": {
                    "type": "string"
                }
            }
        },
...
        "data.UpdateColorRequest": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                }
            }
        },
        "data.UpdateDisplayWindowRequest": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "data.UpdateVisibilityRequest": {
            "type": "object",
            "properties": {
                "is_visible": {
                    "type": "boolean"
                }
            }
        },
...
    }
}